Move MappingMode into rules
