+++ bemuse/src/online/scoreboard-system/ScoreboardClient.ts
@@ -1,4 +1,4 @@
-import { MappingMode } from 'bemuse/app/entities/Options'
+import { MappingMode } from 'bemuse/rules/mapping-mode'
 import type { ScoreCount } from 'bemuse/rules/accuracy'
 
 export interface ScoreboardClient {
